Fix precision problems in the half benchmarks
Need to make sure the comparisons are fair depending on the data type
