+++ benchmark/syclblas/blas1/axpy.cpp
@@ -69,7 +69,8 @@ void run(benchmark::State& state, ExecutorType* executorPtr, index_t size,
   }
 
   std::ostringstream err_stream;
-  if (!utils::compare_vectors(y_temp, y_ref, err_stream, "")) {
+  if (!utils::compare_vectors<data_t, scalar_t>(y_temp, y_ref, err_stream,
+                                                "")) {
     const std::string& err_str = err_stream.str();
     state.SkipWithError(err_str.c_str());
     *success = false;
